requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
python-socketio==5.10.0
python-engineio==4.7.1
eventlet==0.33.3
Werkzeug==3.0.1
email-validator==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
